Klagemodulen: Rename hentVurderinger.ts så det er lettere å lese PR
